backend/requirements.txt
Django==3.0.6
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.4.0
gunicorn==20.0.4
psycopg2-binary
twilio